+++ packages/smarthr-ui/src/components/Balloon/Tooltip/stories/Tooltip.stories.tsx
@@ -1,4 +1,4 @@
-import { FaCircleQuestionIcon } from '../../Icon'
+import { FaCircleQuestionIcon } from '../../../Icon'
 import { Tooltip } from '../Tooltip'
 
 import type { Meta, StoryObj } from '@storybook/react-webpack5'
