+++ packages/smarthr-ui/src/components/Balloon/index.ts
@@ -1,6 +1,2 @@
-/**
- * @deprecated 通常の用途では Tooltip コンポーネントを使用してください。
- * Tour（アプリの初回利用時チュートリアル）のような特殊な用途でのみ使用可能ですが、
- * 将来的には Tour 専用のコンポーネントとして整理される予定です。
- */
 export { Balloon } from './Balloon'
+export { Tooltip } from './Tooltip'
