+++ packages/smarthr-ui/src/components/Dropdown/DropdownTrigger.tsx
@@ -12,7 +12,7 @@ import {
 import { tv } from 'tailwind-variants'
 
 import { tabbable } from '../../libs/tabbable'
-import { Tooltip } from '../Tooltip'
+import { Tooltip } from '../Balloon'
 
 import { DropdownContext } from './Dropdown'
